pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Django project template managed by PDM"
authors = [
    {name = "FAN", email = "admin@migob.digital"},
]
dependencies = [
    "django==4.2.*",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "psycopg>=3.2.4",
    "psycopg2-binary>=2.9.10",
    "markdown>=3.7",
    "django-filter>=24.3",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false
[tool.pdm.scripts]
migrate = "python manage.py migrate"
start = "python manage.py runserver"
post_init = { composite = ["pdm install", "migrate", "start"] }

